Cleanup ReportList layout (#19)
* Cleanup ReportStatusList layout

* Split up ReportStatusList component and cleanup types

---------
